feat(website): digiquant.io de-slop + redesign — honest data, real CTAs, working tabs (#680)
Every fabricated element is gone or relabeled, every checkable number now
matches the repo, and a first-time visitor has something real to do.

Honesty:
- removed the fake price ticker and the 'SIGNALS · LIVE' hardcoded widget;
  the footer now states the page carries no live data and points at Olympus
- fabricated counters fixed: 14->6 strategies (registry on develop),
  42->25 FRED series (macro_series.yaml), 65->90+ watchlist (watchlist.md);
  '12 ms/candle' invented benchmark removed; futures coverage claim dropped
- Kairos consistently labeled PLANNED (diagram node, act kicker, ladder:
  only the backtest rung renders solid); 'human-gated live orders' hero
  claim replaced with what ships today
- 'Managed Atlas' tier no longer sells SLAs for an unbuyable product —
  explicit PLANNED, 'register interest' mailto
- permanently-'deploying' chat embed removed; honest DigiChat pointer kept
- Hermes copy aligned with ADR-0015 (deliberation -> signals, replayable);
  example signal log uses real registry strategy names
- atlas.html: nine->ten phases (matches the ten rendered cards), numbering
  footnote (7/7C/7D split, no phase 8), MIT-only license, stale README table

Design / function:
- hero gains primary CTAs (Olympus dashboard, GitHub quickstart)
- Act V toggles are real tabs now: three panels (dashboard widgets / actual
  MCP tool listing from mcp_server.py / docker quickstart), correct ARIA
  tab semantics, arrow-key navigation — previously clicking only changed an
  outline color
- act compositions varied (left/right rails) to break the 7x-centered
  template monotony; focus-visible styles added; mobile diagram gets
  horizontal scroll with a min legible width
- dead code removed: ticker module + styles, scroll-trigger init that
  matched nothing, data-ohlc pseudo-data attrs, chat-slot CSS; counter and
  draw-in implementations deduplicated into page-motion.js (main.js and
  atlas-main.js carried drifting copies)

Verified in a real browser (Playwright against a local server): zero
console errors on both pages, tab switching confirmed via accessibility
snapshot, full-page screenshot reviewed. make score 10/10/10/10.
